+++ src/js/modal.js
@@ -1,5 +1,5 @@
 //import { renderMovieDetails } from "../index.js";
-import { getImagePathFromTMDB } from '../index.js';
+import { getImagePathFromTMDB, getMovieDetails } from '../index.js';
 const movieGalleryItem = document.querySelector('[data-id]');
 import modalMovieTemplate from '../templates/modal-movie-card.hbs';
 import mainMovieTemplate from '../templates/main-movie-card.hbs';
@@ -114,33 +114,7 @@ if (parsedDataQueue) {
     }
   }
 
-  async function getMovieDetails(movie_id, language = '') {
-    const handler_params = {
-      movie_id: movie_id,
-      language: language,
-    };
-    const URL_handler = new TmdbUrlHandler('TMDB_movieData', handler_params);
-
-    const AxiosMovieParams = {
-      method: 'get',
-      url: URL_handler.toString(),
-    };
-
-    try {
-      const serverResponse = await axios(AxiosMovieParams);
-
-      if (serverResponse.statusText != 'OK' && serverResponse.status != 200) {
-        throw new ServerError(
-          `Unable to get movie data from TMDB. Request: ${AxiosMovieParams.url}. TMDB response: ${serverResponse.statusText}. TMDB RESPONSE STATUS: ${serverResponse.status}`,
-        );
-      }
-      //console.log(serverResponse.data); //debug line
-      return serverResponse.data; //return detailed movie data
-    } catch (error) {
-      console.log(error.message);
-    }
-    return false; //otherwise return false
-  }
+  
 
   async function showMovieDetails(event = new Event('default')) {
     event.preventDefault();
@@ -148,8 +122,7 @@ if (parsedDataQueue) {
     if (event.target.nodeName === 'UL') {
       return;
     }
-    document.body.classList.toggle('modal-open');
-    refs.modalMovie.classList.toggle('backdrop--is-hidden');
+    
     let cardElement = event.target;
     while (cardElement.nodeName != 'LI') {
       cardElement = cardElement.parentNode;
@@ -162,6 +135,8 @@ if (parsedDataQueue) {
     //add movie data to modal window here
     
     renderMovieDetails(movieData);
+    document.body.classList.toggle('modal-open');
+    refs.modalMovie.classList.toggle('backdrop--is-hidden');
   }
   // DEBUG MOVIE DETAILS
   // const testClickOnMovie = { preventDefault() { } }; //dummy Event
